requirements.txt
pyaudio~=0.2.11
flexx~=0.4.2
python-socketio[client]
opencv-python
azure-cognitiveservices-vision-computervision
keras==2.1.5
tensorflow-gpu==1.4
matplotlib
PIL
efficientnet